Support inspecting green threads
Some applications use green threads that have their own stacks. One such
example is Scylla [1] which is using the seastar framework [2], which
provides green threads in the form of seastar::thread [3]. These threads
are created with `setcontext()` and later we switch in/out using
`setjmp()`/`longjmp()`

When debugging Scylla (or any other seastar application) it is essential
to be able to inspect the stacks of these green threads. For this we
used a python extension, which essentially emulates a `longjmp()` call,
restoring the registers from a `jmpbuf`. This approach worked fine for
some time, however we started looking for another way to do this
several reasons:
* Writing to registers is dangerous, any bug in the python script or GDB
  can result in crashing the debugged live application. Very undesirable
  when having to debug on a production server.
* Doesn't work in coredumps, where registers are not writable.
* Stopped working for some time, resulting in a crash of GDB.

This patch aims to solve this problem by providing a family of commands
which allow inspecting green threads.
* fiber view - allows inspecting a green thread. This command takes a
  list of registers as its arguments. More on these later.
* fiber reset - reset the viewed fiber to the stack proper.

The implementation turned out to be quite simple. The command just
propagates the registers to override to the sentinel frame, which will
return the overridden values for registers of interest, instead of
consulting the current register cache, thus simulating a register
restore. This has two important advantages: it works in coredumps and it
is save. If the user passes wrong register values, it might crash gdb,
but as the application state is not mangled with, it should survive
unscathed.

TODO:
* Better interface for the command: instead of a fixed set of registers
  (those that happen to be in glibc's jmpbuf on amd64), allow the user
  to pass any amount of named registers to override.
* More documentation.
* More testing.
* Code style.

I apologize for the mixed code style, but I wanted to probe my approach
first, before spending more time on polishing.